Fix for texture animation
When a texture animation has multiple rows, like the explosion boomer, it was cycling through the frames in the wrong order. This fix starts with the top row and moves down, instead of starting with the bottom row and moving up. Tile 0 is now the top left, rather than the bottom left.